Fix background bleeding issue.
This is a fix for an issue that sometimes happened where background
color seemed to bleed through between map tiles. The reason this
happens seemed to be that the texture is sampled a bit outside the
tile area specified, sampling transparent values.

The fix offsets texture coordinates by a tiny bit (0.001 at the
moment) towards the inside of the tile.