Add "base_y" to objects.
Add a field "base_y" to objects which determine the y coordinate at
which they "touch the ground". This is used for sorting the objects
based on the location where they are actually supposed to be on the
ground.

The "base_y" value is a ratio in range [0.0, 1.0]. The real value is
calculated from multiplying it with the object height.